+++ custom/justech_l10n_do_ncf/wizards/__init__.py
@@ -1,3 +1,4 @@
 from . import fiscal_diagnostic_wizard
 from . import ncf_migration_wizard
 from . import ncf_reconcile_wizard
+from . import ncf_void_wizard